port prove_namespace test to advz/namespaces.rs
