+++ primitives/src/vid/advz.rs
@@ -615,7 +615,6 @@ mod tests {
         rand::{CryptoRng, RngCore},
         vec,
     };
-    use digest::{generic_array::ArrayLength, OutputSizeUser};
     use sha2::Sha256;
 
     #[test]
@@ -799,71 +798,12 @@ mod tests {
         }
     }
 
-    fn prove_namespace_generic<E, H>()
-    where
-        E: Pairing,
-        H: Digest + DynDigest + Default + Clone + Write,
-        <<H as OutputSizeUser>::OutputSize as ArrayLength<u8>>::ArrayType: Copy,
-    {
-        // play with these items
-        let (payload_chunk_size, num_storage_nodes) = (4, 6);
-        let num_polys = 4;
-
-        // more items as a function of the above
-        let payload_elems_len = num_polys * payload_chunk_size;
-        let payload_bytes_len = payload_elems_len * modulus_byte_len::<E>();
-        let mut rng = jf_utils::test_rng();
-        let payload_bytes = init_random_payload(payload_bytes_len, &mut rng);
-        let srs = init_srs(payload_elems_len, &mut rng);
-
-        let advz = Advz::<E, H>::new(payload_chunk_size, num_storage_nodes, srs).unwrap();
-        let d = advz.disperse(&payload_bytes).unwrap();
-
-        // TEST: verify "namespaces" (each namespace is a polynomial)
-        // This test is currently trivial: we simply repeat the commit computation.
-        // In the future there will be a proper API that can be tested meaningfully.
-
-        // encode payload as field elements, partition into polynomials, compute
-        // commitments, compare against VID common data
-        let elems_iter = bytes_to_field::<_, E::ScalarField>(payload_bytes);
-        for (coeffs_iter, poly_commit) in elems_iter
-            .chunks(payload_chunk_size)
-            .into_iter()
-            .zip(d.common.poly_commits.iter())
-        {
-            let mut coeffs: Vec<_> = coeffs_iter.collect();
-            advz.eval_domain.fft_in_place(&mut coeffs);
-
-            let poly = <UnivariateKzgPCS::<E> as PolynomialCommitmentScheme>::Polynomial::from_coefficients_vec(coeffs);
-            let my_poly_commit = UnivariateKzgPCS::<E>::commit(&advz.ck, &poly).unwrap();
-            assert_eq!(my_poly_commit, *poly_commit);
-        }
-
-        // compute payload commitment and verify
-        let commit = {
-            let mut hasher = H::new();
-            for poly_commit in d.common.poly_commits.iter() {
-                // TODO compiler bug? `as` should not be needed here!
-                (poly_commit as &<UnivariateKzgPCS<E> as PolynomialCommitmentScheme>::Commitment)
-                    .serialize_uncompressed(&mut hasher)
-                    .unwrap();
-            }
-            hasher.finalize()
-        };
-        assert_eq!(commit, d.commit);
-    }
-
-    #[test]
-    fn prove_namespace() {
-        prove_namespace_generic::<Bls12_381, Sha256>();
-    }
-
     /// Routine initialization tasks.
     ///
     /// Returns the following tuple:
     /// 1. An initialized [`Advz`] instance.
     /// 2. A `Vec<u8>` filled with random bytes.
-    fn avdz_init() -> (Advz<Bls12_381, Sha256>, Payload) {
+    pub(super) fn avdz_init() -> (Advz<Bls12_381, Sha256>, Payload) {
         let (payload_chunk_size, num_storage_nodes) = (4, 6);
         let mut rng = jf_utils::test_rng();
         let srs = init_srs(payload_chunk_size, &mut rng);
@@ -873,11 +813,11 @@ mod tests {
     }
 
     /// Convenience wrapper to assert [`VidError::Argument`] return value.
-    fn assert_arg_err<T>(res: VidResult<T>, msg: &str) {
+    pub(super) fn assert_arg_err<T>(res: VidResult<T>, msg: &str) {
         assert!(matches!(res, Err(Argument(_))), "{}", msg);
     }
 
-    fn init_random_payload<R>(len: usize, rng: &mut R) -> Payload
+    pub(super) fn init_random_payload<R>(len: usize, rng: &mut R) -> Payload
     where
         R: RngCore + CryptoRng,
     {
@@ -886,7 +826,7 @@ mod tests {
         Payload::from_vec(bytes_random)
     }
 
-    fn init_srs<E, R>(num_coeffs: usize, rng: &mut R) -> UnivariateUniversalParams<E>
+    pub(super) fn init_srs<E, R>(num_coeffs: usize, rng: &mut R) -> UnivariateUniversalParams<E>
     where
         E: Pairing,
         R: RngCore + CryptoRng,
@@ -895,7 +835,7 @@ mod tests {
             .unwrap()
     }
 
-    fn modulus_byte_len<E>() -> usize
+    pub(super) fn modulus_byte_len<E>() -> usize
     where
         E: Pairing,
     {
